# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an order for published courses",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}/payment-proof": {
      "post": {
        "operationId": "OrdersController_submitPaymentProof",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPaymentProofDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "proofId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Submit a payment proof for an own order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}/verify": {
      "patch": {
        "operationId": "OrdersController_verifyOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Verify an order payment (admin)",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}/activate": {
      "post": {
        "operationId": "OrdersController_activateOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderActivationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Activate enrollments from a paid order (admin)",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Cancel a pending order (admin)",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/projects": {
      "get": {
        "operationId": "ProjectsController_listPublished",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectCardDto"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List published projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{id}/media": {
      "post": {
        "operationId": "ProjectMediaController_attachMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachProjectMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMediaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Attach media to a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/media/{mediaId}": {
      "delete": {
        "operationId": "ProjectMediaController_detachMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Detach media from a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/members": {
      "post": {
        "operationId": "ProjectMembersController_assignMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProjectMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projectId": {
                      "type": "string"
                    },
                    "memberId": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Assign a collective member to a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/members/{memberId}": {
      "delete": {
        "operationId": "ProjectMembersController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Remove a member assignment from a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xolvon Backend API",
    "description": "HTTP contract for the Xolvon backend (global prefix /api). Auth: `Authorization: Bearer <accessToken>` (JWT, issued by POST /api/auth/register and POST /api/auth/login). Public routes need no credentials; routes listed with bearer security reject missing or invalid tokens with 401.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token. Send it as `Authorization: Bearer <accessToken>`."
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "courseIds": {
            "description": "Course ids to purchase (UUID v4, unique, non-empty).",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "description": "Order notes, max 500 characters, sanitized on write.",
            "maxLength": 500
          }
        },
        "required": [
          "courseIds"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Order item id (UUID v4)."
          },
          "courseId": {
            "type": "string",
            "description": "Course id (UUID v4)."
          },
          "price": {
            "type": "number",
            "description": "Item price in IDR (server-computed)."
          }
        },
        "required": [
          "id",
          "courseId",
          "price"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Order id (UUID v4)."
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "cancelled"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Total amount in IDR (server-computed)."
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp."
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601 timestamp."
          },
          "verifiedAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp, null until verified."
          }
        },
        "required": [
          "id",
          "status",
          "amount",
          "notes",
          "items",
          "createdAt",
          "updatedAt",
          "verifiedAt"
        ]
      },
      "SubmitPaymentProofDto": {
        "type": "object",
        "properties": {
          "objectKey": {
            "type": "string",
            "description": "Object key of the uploaded payment proof, max 1000 characters.",
            "maxLength": 1000
          }
        },
        "required": [
          "objectKey"
        ]
      },
      "OrderActivationResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "orderId": {
            "type": "string",
            "description": "Order id (UUID v4)."
          },
          "activatedCoursesCount": {
            "type": "number",
            "description": "Number of enrollments activated by the call."
          }
        },
        "required": [
          "message",
          "orderId",
          "activatedCoursesCount"
        ]
      },
...
      "ProjectCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Project id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "published"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "type",
          "summary",
          "status"
        ]
      },
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Project title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 200 characters.",
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "description": "Controlled project type string, max 100 characters (not a final enum).",
            "maxLength": 100
          },
          "summary": {
            "type": "string",
            "description": "One-line summary, max 500 characters, sanitized on write.",
            "maxLength": 500
          },
          "problem": {
            "type": "string",
            "description": "Story section \"Problem\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          },
          "solution": {
            "type": "string",
            "description": "Story section \"Solution\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          },
          "techStack": {
            "type": "string",
            "description": "Raw tech-stack text (parsed at read boundary), max 1000 characters.",
            "maxLength": 1000
          },
          "result": {
            "type": "string",
            "description": "Story section \"Result\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          }
        },
        "required": [
          "title",
          "slug"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "AttachProjectMediaDto": {
        "type": "object",
        "properties": {
          "objectKey": {
            "type": "string",
            "description": "Object key of the uploaded file, max 500 characters.",
            "maxLength": 500
          },
          "mediaType": {
            "type": "string",
            "description": "Controlled media type string, max 50 characters (not a final enum).",
            "maxLength": 50
          },
          "sortOrder": {
            "type": "number",
            "description": "Display order, min 0.",
            "minimum": 0
          }
        },
        "required": [
          "objectKey",
          "mediaType"
        ]
      },
      "ProjectMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Media row id (UUID v4)."
          },
          "projectId": {
            "type": "string",
            "description": "Owning project id (UUID v4)."
          },
          "objectKey": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "projectId",
          "objectKey",
          "mediaType",
          "sortOrder"
        ]
      },
      "AssignProjectMemberDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "description": "Collective member id (UUID v4)."
          },
          "role": {
            "type": "string",
            "description": "Explicit role attribution, max 100 characters (controlled string, not a final enum).",
            "maxLength": 100
          }
        },
        "required": [
          "memberId",
          "role"
        ]
      },
...
    }
  }
}